Fix void * to integer casts for 64-bit Intel systems (STR #1446)
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@5499 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ test/input.cxx
@@ -61,7 +61,7 @@ void button_cb(Fl_Widget *,void *) {
 
 void color_cb(Fl_Widget* button, void* v) {
   Fl_Color c;
-  switch ((int)v) {
+  switch ((long)v) {
   case 0: c = FL_BACKGROUND2_COLOR; break;
   case 1: c = FL_SELECTION_COLOR; break;
   default: c = FL_FOREGROUND_COLOR; break;
